+++ config/services/ui.mjs
@@ -1,4 +1,4 @@
-import { generateTraefikConfig } from './index.mjs'
+import { generateTraefikConfig, getContainerTagSuffix } from './index.mjs'
 
 /*
  * Export a single method that resolves the service configuration
@@ -22,7 +22,7 @@ export const resolveServiceConfiguration = ({ utils }) => {
       // Image to run (different in dev)
       image: PROD ? 'itsmorio/ui' : 'devmorio/ui',
       // Image tag (version) to run
-      tag: utils.getPreset('MORIO_VERSION_TAG'),
+      tag: utils.getPreset('MORIO_VERSION_TAG') + getContainerTagSuffix(utils),
       // Don't attach to the default network
       networks: { default: null },
       // Instead, attach to the morio network